KQL query is incorrect
New users just copy paste and don't know the name of the actual field.


url_has_any	dynamic	Filter only Web sessions for which the URL field has any of the values listed. If specified, and the session is not a web session, no result will be returned. The length of the list is limited to 10,000 items.